Moved stop screen logic into new controller and added stop screen url
