+++ src/controllers/confirm.company.controller.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, Response } from "express";
 import { Templates } from "../types/template.paths";
 import { CompanyProfile } from "@companieshouse/api-sdk-node/dist/services/company-profile/types";
 import { Session } from "@companieshouse/node-session-handler";
-import { COMPANY_LOOKUP, CREATE_TRANSACTION_PATH} from "../types/page.urls";
+import { COMPANY_LOOKUP, CREATE_TRANSACTION_PATH, SHOW_STOP_PAGE_PATH, URL_QUERY_PARAM} from "../types/page.urls";
 import { urlUtils } from "../utils/url";
 import { getCompanyProfile } from "../services/company.profile.service";
 import { buildAddress, formatForDisplay } from "../services/confirm.company.service";
@@ -14,11 +14,8 @@ export const get = async (req: Request, res: Response, next: NextFunction) => {
     const companyNumber = req.query.companyNumber as string;
     const companyProfile: CompanyProfile = await getCompanyProfile(companyNumber);
 
-    const transactionId = "No transaction yet";
-
     if (await currentOrFutureDissolved(session, companyNumber)){
-      const content = setContent(companyProfile)
-      displayStopPage(res, content)
+      res.redirect(urlUtils.setQueryParam(SHOW_STOP_PAGE_PATH, URL_QUERY_PARAM.COMPANY_NUM, companyNumber));
     }
     else {
       const pageOptions = await buildPageOptions(session, companyProfile);
@@ -29,27 +26,6 @@ export const get = async (req: Request, res: Response, next: NextFunction) => {
     return next(e);
   }
 };
-
-const setContent = (companyProfile: CompanyProfile) => {
-  if (companyProfile.companyName == undefined || companyProfile.companyName == "") {
-    companyProfile.companyName = "This company"
-  }
-  
-  const content = {
-    pageHeader: "Company is Dissolved or it's in the process of being dissolved",
-    pageBody: `
-    <p>${companyProfile.companyName} cannot use this service because it has been dissolved, or it's in the process of being dissolved.</p>
-
-    <p><a href="https://www.gov.uk/guidance/company-restoration-guide" data-event-id="read-the-company-restoration-guide-link">Read the Company Restoration Guide</a> to find out more about restoring a company name to the register.</p>
-    <p>If this is the wrong company, <a href="/officer-filing-web" data-event-id="start-the-service-again-link">start the service again</a>.</p>
-    <p><a href="https://www.gov.uk/contact-companies-house" data-event-id="contact-us-link">Contact us</a> if you have any questions.</p>
-    `
-  }
-
-  return{
-    content
-  }
-}
 
 const buildPageOptions = async (session: Session, companyProfile: CompanyProfile): Promise<Object> => {
   companyProfile = formatForDisplay(companyProfile);
@@ -91,8 +67,4 @@ export const post = async (req: Request, res: Response, next: NextFunction) => {
 
 const createNewOfficerFiling = async (session: Session) => {
     const transactionId: string = "";
-};
-
-const displayStopPage = (res: Response, content: { content: {pageHeader: string, pageBody: string}}) => {
-  return res.render(Templates.STOP_PAGE, content);
 };
